move styles to other file
