+++ App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import StartMenu from './components/startMenu';
+import StartMenu from './components/startMenu/startMenu';
 
 export default function App() {
   return (
